+++ README.md
@@ -27,11 +27,11 @@ fullstackopen.com course exercises
   - [x] 1.14*: anecdotes step3 [[instructions][link3]]
 
 - [ ] [part2][part2] - Communicating with server ![in progress][progress]
-  - [ ] 2.1: course contents step6 [[instructions][link4]]
-  - [ ] 2.2: Course contents step7 [[instructions][link4]]
-  - [ ] 2.3*: Course contents step8 [[instructions][link4]]
-  - [ ] 2.4: Course contents step9 [[instructions][link4]]
-  - [ ] 2.5: separate module [[instructions][link4]]
+  - [x] 2.1: course contents step6 [[instructions][link4]]
+  - [x] 2.2: Course contents step7 [[instructions][link4]]
+  - [x] 2.3*: Course contents step8 [[instructions][link4]]
+  - [x] 2.4: Course contents step9 [[instructions][link4]]
+  - [x] 2.5: separate module [[instructions][link4]]
   - [ ] 2.6: The Phonebook Step1 [[instructions][link5]]
   - [ ] 2.7: The Phonebook Step2 [[instructions][link5]]
   - [ ] 2.8: The Phonebook Step3 [[instructions][link5]]
